test: update tests

+++ driver/anchor_tx_constructor/anchor_tx_constructor_test.go
@@ -36,7 +36,7 @@ func (s *AnchorTxConstructorTestSuite) SetupTest() {
 }
 
 func (s *AnchorTxConstructorTestSuite) TestGasLimit() {
-	s.Greater(AnchorGasLimit, uint64(0))
+	s.Greater(AnchorGasLimit, 0)
 }
 
 func (s *AnchorTxConstructorTestSuite) TestAssembleAnchorTx() {
